pyproject.toml
[tool.poetry]
name = "duplicate_images"
version = "0.2.0"
description = ""
authors = ["Lene Preuss <lene.preuss@gmail.com>"]
repository = "https://gitlab.com/lilacashes/DuplicateImages.git"
homepage = "https://gitlab.com/lilacashes/DuplicateImages"
readme = "README.md"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Utilities"
]

[tool.poetry.dependencies]
python = "^3.8.5"
Wand = "^0.6.5"
pillow = "^8.1.0"
imagehash = "^4.2.0"

[tool.poetry.dev-dependencies]
lz4 = "^3.1.1"
mypy = "^0.790"
flake8 = "^3.8.4"
Wand = "^0.6.5"
pytest = "^6.2.1"
pylint = "^2.6.0"
filprofiler = "^0.14.0"

[tool.poetry.scripts]
find-dups = "duplicate_images.duplicate:main"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
